+++ lucia/src/dnsn.rs
@@ -3,9 +3,13 @@
 //! `lucia` abstracts different serialization/deserialization frameworks to enhance de-coupling,
 //! enable choice and improve experimentation.
 
+#[cfg(feature = "borsh")]
+mod borsh;
 mod deserialize;
 #[cfg(feature = "miniserde")]
 mod miniserde;
+#[cfg(feature = "rkyv")]
+mod rkyv;
 #[cfg(feature = "serde_json")]
 mod serde_json;
 #[cfg(feature = "serde-xml-rs")]
@@ -14,8 +18,12 @@ mod serde_xml_rs;
 mod serde_yaml;
 mod serialize;
 
+#[cfg(feature = "borsh")]
+pub use self::borsh::*;
 #[cfg(feature = "miniserde")]
 pub use self::miniserde::*;
+#[cfg(feature = "rkyv")]
+pub use self::rkyv::*;
 #[cfg(feature = "serde_json")]
 pub use self::serde_json::*;
 #[cfg(feature = "serde-xml-rs")]