+++ CHANGELOG.md
@@ -6,4 +6,6 @@
 
 * More documentation, tests and endpoints. [#11](https://github.com/c410-f3r/lucia/pull/11)
 
-* Add related API types. [#16](https://github.com/c410-f3r/lucia/pull/11)
+* Add related API types. [#16](https://github.com/c410-f3r/lucia/pull/16)
+
+* Add support for borsh and rkyv. [#23](https://github.com/c410-f3r/lucia/pull/23)